refactor(#192): consolidate TypeScript types into src/types
- Create src/types/index.ts as central export point for all types
- Move CampaignStatus, CampaignInfo, CampaignStats types to types/soroban.ts
- Add new types: PlatformConfig, StatusVariant, ContributionRecord, InitializeParams
- Move SignFn and ContractError to types/contract.ts with proper exports
- Expand Campaign interface to include all CampaignInfo fields
- Update src/lib/soroban.ts to import and re-export types from types/
- Update src/lib/contract.ts to import from types/contract.ts
- Remove duplicate type definitions scattered across the codebase

Closes #192
